리뷰리스트 가져오는 도메인에서 리뷰 리스트 가져오기 및 리뷰 상세보기 화면 넘어가기 기능 적용

+++ Moari/CategoryDetail/Actor/Entity/ReviewModel.swift
@@ -12,9 +12,9 @@ struct Review {
     var reviewCategory: Int!
     var title: String!
     var imageUrl: String!
-    var starRate: Float!
+    var starRate: Double!
     
-    init(reviewId: Int, reviewCategory: Int, title: String, imageUrl: String, starRate: Float) {
+    init(reviewId: Int, reviewCategory: Int, title: String, imageUrl: String, starRate: Double) {
         self.reviewId = reviewId
         self.reviewCategory = reviewCategory
         self.title = title
